move backpack-related files to sharedComponents
