+++ apps/src/sharedComponents/backpack/Backpack.jsx
@@ -11,10 +11,11 @@ import {makeEnum} from '@cdo/apps/utils';
 import javalabMsg from '@cdo/javalab/locale';
 import msg from '@cdo/locale';
 
+import {DisplayTheme} from '../../javalab/DisplayTheme';
+import JavalabButton from '../../javalab/JavalabButton';
+import JavalabDialog from '../../javalab/JavalabDialog';
+
 import {BackpackAPIContext} from './BackpackAPIContext';
-import {DisplayTheme} from './DisplayTheme';
-import JavalabButton from './JavalabButton';
-import JavalabDialog from './JavalabDialog';
 
 import moduleStyles from './backpack.module.scss';
